add encryption_files module
