+++ lib/enigma.rb
@@ -1,8 +1,10 @@
 require './lib/encryption'
+require './lib/encryption_files'
 require 'date'
 require 'pry'
 
 class Enigma
+  include EncryptionFiles
   include Encryption
   def initialize
     @alphabet = ("a".."z").to_a << " "
@@ -33,36 +35,4 @@ class Enigma
     }
   end
 
-  def encrypt_txt
-    txt_file_to_encrypt = ARGV[0]
-    encrypted_new_file_name = ARGV[1]
-
-
-    file = File.open(txt_file_to_encrypt, "r")
-    message = file.read.chomp
-    encrypted_message = encrypt(message)
-
-    new_file = File.open(encrypted_new_file_name, "w")
-    new_file.write(encrypted_message[:encryption])
-    new_file.close
-
-    puts "Created \'#{encrypted_new_file_name}\' with the key #{encrypted_message[:key]} and date #{encrypted_message[:date]}"
-  end
-
-  def decrypt_txt
-    txt_file_to_decrypt = ARGV[0]
-    new_decrypted_file_name = ARGV[1]
-    input_key = ARGV[2]
-    input_date = ARGV[3]
-
-    file = File.open(txt_file_to_decrypt, "r")
-    message = file.read.chomp
-    decrypted_message = decrypt(message, input_key, input_date)
-
-    new_file = File.open(new_decrypted_file_name, "w")
-    new_file.write(decrypted_message[:decryption])
-    new_file.close
-
-    puts "Created \'#{new_decrypted_file_name}\' with the key #{decrypted_message[:key]} and date #{decrypted_message[:date]}"
-  end
 end
