+++ spec/spec_helper.rb
@@ -1,0 +1,2 @@
+# require 'simplecov'
+# SimpleCov.start
